app/requirements.txt
prettytable
requests
python-dotenv
